+++ app/src/main/java/com/example/healthcaresystem/api/RetrofitInstance.kt
@@ -5,7 +5,7 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 object RetrofitInstance {
-    private const val BASE_URL = "http://192.168.0.104:4000"
+    private const val BASE_URL = "http://192.168.1.50:3000"
 
     private val client = OkHttpClient.Builder()
         .build()
