+++ app/src/main/java/com/example/healthcaresystem/user/Intro2.kt
@@ -1,4 +1,4 @@
-package com.example.healthcaresystem.User
+package com.example.healthcaresystem.user
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,20 +9,21 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.example.healthcaresystem.R
 
-class SignUpSuccess : AppCompatActivity() {
+class Intro2 : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_signup_success)
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.success)) { v, insets ->
+        setContentView(R.layout.activity_intro2)
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.intro2)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
         }
-        val nextButton = findViewById<TextView>(R.id.btnSignIn)
+        val nextButton = findViewById<TextView>(R.id.btnNext)
         nextButton.setOnClickListener {
-            val intent = Intent(this, SignIn::class.java)
+            val intent = Intent(this, Intro3::class.java)
             startActivity(intent) // Chuyển đến SecondActivity
         }
+
     }
 }
