+++ app/src/main/java/com/example/healthcaresystem/user/home/HomeActivity.kt
@@ -1,4 +1,4 @@
-package com.example.healthcaresystem.User.home
+package com.example.healthcaresystem.user.home
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -15,7 +15,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHost
 import com.example.healthcaresystem.ui.theme.HealthCareSystemTheme
 
 class HomeActivity : ComponentActivity() {
