+++ app/src/main/java/com/example/healthcaresystem/user/home/ScreenInfo.kt
@@ -1,4 +1,4 @@
-package com.example.healthcaresystem.User.home
+package com.example.healthcaresystem.user.home
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
