+++ app/src/main/java/com/example/healthcaresystem/user/home/model/Doctor.kt
@@ -1,4 +1,4 @@
-package com.example.healthcaresystem.User.home.model
+package com.example.healthcaresystem.user.home.model
 
 // Dữ liệu bác sĩ
 data class Doctor(
