+++ app/src/main/java/com/example/healthcaresystem/user/home/model/FAQItem.kt
@@ -1,4 +1,4 @@
-package com.example.healthcaresystem.User.home.model
+package com.example.healthcaresystem.user.home.model
 
 data class FAQItem(
     val question: String,
