+++ app/src/main/res/xml/network_security_config.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <network-security-config xmlns:android="http://schemas.android.com/apk/res/android">
     <domain-config cleartextTrafficPermitted="true">
-        <domain includeSubdomains="true">192.168.0.104</domain>
+        <domain includeSubdomains="true">192.168.1.50</domain>
 <!--        //May cua khoa 192.168.0.104-->
     </domain-config>
 </network-security-config>
